indent/xml.vim: fix indentation when tags start with colon or underscore
The XML spec says that element names can start with some non-alphabetic
characters: https://www.w3.org/TR/REC-xml/#NT-NameStartChar

However when indenting XML containing elements whose name start with
a colon or an underscore vim fails to get the indentation levels.

So, allow colons and underscores in b:xml_indent_open.

In order to do that, the \a atom needs to be expanded to its equivalent
characters class before adding ':' and '_' to it because character class
cannot contain atoms (excepts some special ones), see the NOTE at the
end of "help \[", (i.e.  [:_\a] would not have been valid).