Include xml indent Script
The indent script seems unmaintained and since I am already maintaining
the xml filetype and xmlformat script, I can also start maintaining the
corresponding xml indent script.

This means, fixing a bug, that the xml comments are not correctly
indented. Previously, it was done on purpose to not indent comments, but
that would confuse indenting of the following lines and also caused
strange behaviour like https://github.com/vim/vim/issues/3200

So remove that part entirely. Courtesy E Kawashima for noticing and
fixing.