syntax: import dtd syntax script
The syntax/xml.vim syntax script is requires the dtd.vim syntax script,
so add it here as well. It was previously also maintained by Johannes
Zellner, so move him to the previous maintainer line. Also while at it,
slightly re-order the header lines to match those of the other runtime
files.