xmlformat: Reformat tag content according to textwidth
the xmlformat script did not check the textwidth setting of a buffer.
For the tags, this shouldn't be a problem, as those are usually much
shorter than the tag content.

However overlong tag content would not be wrapped correctly, rather it
would format the content as one long line.

Therefore, if textwidth setting is set, look for whitespace a bit for
the textwidth limit (currently 10 chars) and wrap it and add indent to
it, if required. If textwidth is not set, take the content as is.

Add a test to verify the behaviour

closes #18