indent: behave sanely for docbk filetype
fixes #20

Also include a very simple docbk indent test. The result is not really
satisfactory, so the test shows that the indent script might need more
adjustments, as there still are some staircase issues with the resulting
indented xml docbk file. But that can be adjusted later if needed.

The problem mentioned in #20 should however be fixed:

So adding new text like this:
 <varlistentry>
  <term></term>
   <listitem>
    <para>
     </para>
      </listitem>
       </varlistentry>
        </variablelist>

does not cause the staircase indication anymore