indent: Fix recently broken changes regarding comment and non-xml indent
With 'ptagopen' we are now also matching newlines in search for the end
of the start tags. I do not know if there are any performance issues or
the like with this implementation.

All tests pass and I have updated one to match the correct indentation
of comments after multi-line tags. I also added an extra open tag with
dashes in its name since it is valid XML.