indent: fix erroneous indent for empty closing tags
fixes #25

This is a regression from the initial XML indent script, in that the
XmlIndentSum() function would take a line such as
  foobar<tag/> and indent it by one additional shiftwidth, resulting in
a staircase like indentation if several of those lines followed.

Fix this by also testing, whether a line matches <tag/>\s*$ and so that
this condition would not match anymore.

Add a test, so that this does not regress anymore